Update Renderpath
Looks for Renderoutputs that have a path defined and updates it to the correct value dependent on the scene's name.

"Having a path defined" means that it literally just needs a string in the filename field.

I usually put a generic "path" into any filename field for Renderoutputs that I save as presets.

Signed-off-by: Alexander Kucera <a.kucera@babylondreams.de>