AOVs are now output into a subdirectory
All AOVs get a folder with the name of the AOV except rgba, which get the root folder as output path.

This was necessary because we are running into problems with huge numbers of files in one folder (about 45Gb and 70K files for one shot in this specific case).

I am still not sure how to treat passes yet. At the moment they land in the respective AOV's folder with an <_passname> in the filename.

Ideally I'd like to get to a point where each pass gets a subfolder on its own and in there are folders for AOVs.

Signed-off-by: Alexander Kucera <a.kucera@babylondreams.de>