Made a mesh of Git
This is my hamfisted way to work around a commit error I made and that I had no idea how to solve otherwise. I tried a few other things (the commits right before this) but that didn't have the intended effect. So now I simply copied the file I wanted to have back from an earlier commit and pasted it as a change for this commit.

TL;DR;:
This restores the AOV subfolder feature to the develop branch, while keeping the master branch at the old "no subfolders" stage. I need that old way of working until the current project is finished.

Signed-off-by: Alexander Kucera <a.kucera@babylondreams.de>